add include specs
This patch introduces a notion of include specs to allow
for finer grained control over what data is loaded.

The `pggen` root module grows a new submodule `include` in
this patch which defines the data structure and DSL syntax
for include specs.

The generated code now produces two public methods instead
of three, with the most flexible being `<Entity>FillIncludes`

The documentation included in this patch explains the concept
further.

Closes #8

+++ pggen/test/cli_test.go
@@ -108,9 +108,17 @@ name = "small_entities"
 }
 
 func TestCLI(t *testing.T) {
+	debug := false
+	debugEnvVar := os.Getenv("PGGEN_DEBUG_CLI")
+	if debugEnvVar == "1" || debugEnvVar == "true" {
+		debug = true
+	}
+
 	testDir, err := ioutil.TempDir("", "pggen_cli_test")
 	chkErr(t, err)
-	defer os.RemoveAll(testDir)
+	if !debug {
+		defer os.RemoveAll(testDir)
+	}
 
 	repoRoot, err := getRepoRoot()
 	chkErr(t, err)
@@ -120,6 +128,10 @@ func TestCLI(t *testing.T) {
 
 	// build the executable we are going to be testing
 	cmd := exec.Command("go", "build", "-o", exe, mainSrc)
+	if debug {
+		cmd = exec.Command(
+			"go", "build", "-gcflags", "all=-N -l", "-o", exe, mainSrc)
+	}
 	err = cmd.Run()
 	chkErr(t, err)
